Set ActionCable's allowed_request_origins to the machine's IPs during development
By default when in development, Action Cable's allowed_request_origins
includes http://localhost:3000.  Because it is compelling to experiment
with Action Cable applications using multiple machines, when running in
development this update finds all IPv4 addresses and adds them into
allowed_request_origins.  It does not affect test and production
environments at all.

When running "rails s -b 0.0.0.0" the system will allow other machines
on the same network to easily use Action Cable channels without having
to explicitly configure a static IP address in development.rb.

As a convenience for testing, this new helper file is included:

    railties/test/server_helpers.rb

It allows objects built with the Rails::Server class to go through
the .app.initialize! process, and also be torn back down enough to
have new settings applied for the host, port, and environment and
then be initialized again. (This gets used here to confirm that there
are no changes to allowed_request_origins when in production.)