Push :defaults extraction down one level
Since e852daa6976cc6b6b28ad0c80a188c06e226df3c only the verb methods
where extracting the defaults options. It was merged a fix for the
`root` method in 31fbbb7faccba25b2e3b5e10b8fca1468579d629 but `match`
was still broken since `:defaults` where not extracted.

This was causing routes defined using `match` and having the `:defaults`
keys to not be recognized.

To fix this it was extracted a new private method with the actual
content of `match` and the `:defaults` extracting was moved to `match`.