Add Enumerable#single
This is inspired by other languages and frameworks, such as LINQ, which
has very big distinction between `first` and `single` element of the
collection.

- `first` normally returns the top element, and the developer assumes
  there could be many;
- `single` returns one and only one element, and it is an error if there
  are none or more than one.

We, in Ruby world, very often write `Account.where(name: 'foo').first`
assuming there's only one element that can be returned there.

But in majority of the cases, we really want a `single` element.

The problems with using `first` in this case:
- developer needs to explicitly double check the result isn't `nil`
- in case of corrupted data (more than one item returned), it will never
  be noticed

`Enumerable#single` addresses those problems in a very strong and
specific way that may save the world by simply switching from `first` to
`single`.

Also, we have the answer to universe - `fourty_two`, may as well save the
world with a `single` before waiting for that answer :)