Translate PG::NotNullViolation to AR::NotNull
We'd like to be able to better present errors that occur when a record
passes in-memory validation but fails to insert or update to the
database due to a null constraint on a column. Introduce
ActiveRecord::NotNull as a wrapper around these database errors, and
raise this error type when the Postgres adapter raises a
PG::NotNullViolation. Since we'll need to know what column caused the
issue to correctly display the error, we parse the PG::NotNullViolation
message to extract the column.

In applications developers generally see PG::NotNullViolations in their
logs or bug trackers when they've set a null constraint on a column but
have not defined a corresponding presence validation on the model. This
often manifests as a 500 response to application users when they've not
populated a specific column, but isn't presented in a good way as it
would be if the validation were present.

While this change does not itself solve that problem, it does provide
the start of a standard interface we can use internally to rescue and
handle varying errors caused by adapters when a null violation occurs.

Wrap the PG::NotNullViolation in an ActiveRecord error with knowledge of
the column which caused the error.