Translate NotNull errors to RecordInvalid
NotNull provides the standard interface that connection adapters can use
to inform us of not null constraint violations. This allows us to rescue
them in order to provide a familiar and useful interface to invalid
records: the ActiveRecord::RecordInvalid error combined with populating
`errors[column]` with a "blank" error as if the model had defined the
presence validation implied by the null constraint.

Rescue ActiveRecord::NotNull, populate errors[column] with an
appropriate `:blank` message, and raise RecordInvalid.