Clarify ActiveSupport::Rescuable#rescue_from behaviour to fallback to cause
`ActiveSupport::Rescuable.rescue_from` will select handlers by first
finding a handler for the specific exception. If no handler is found
and the exception has a `cause`, it will search a handler for the
cause and so on.

There are two ways to interpret `rescue_from`. The first one is to
assume that the block argument will be one of the specified classes.
The second is to assume that children to the specified classes will be
rescued by their parent if no handler is found.

Here is a quick example:

```ruby
rescue_from ActiveResource::ClientError do |error|
  if error.response.code == "402"
    render :payment_required
  else
    render :error
  end
end

def fetch_remote_resource
  main_resource = RemoteResource.find
  # load more resources
rescue ActiveResource::ResourceNotFound => e
  log("failed to load associations of #{main_resource.id}")
end
```

If the load of `main_resource` fails, the `rescue_from` will be called
with a `NoMethodError`. `NoMethodError` obviously doesn't respond to
`response.code`. I understand the rationale of matching
`exception.cause` in cases where the handler is very generic, like
logging or rendering an error message. For specific handlers, it means
the handler needs to check if the handler is exception is actually
what's expected. In the previous example, I think it would have been
sane to let the `NoMethodError` bubble up to the middleware and get
caught by `ActionDispatch::DebugExceptions` or equivalent.