Don't bust AR query cache when types are not same object
AR's query cache currently depends on the type objects being not only
equivalent but also to be the same object. Either reloading the type
information or using custom type objects that aren't cached will
unnecessarily bust the cache.

This is fixed in 5.0 as an ancillary part of 574f255629a45cd67babcfb9bb8e163e091a53b8
but here I also add a test for the condition.