Avoid `RangeError` without explicit `limit`
Raising `RangeError` was introduced in order to ease to handle out of
range numbers for `find` and `find_by` in postgresql adapter with
prepared statements since #17459. At that time, `Type::Integer` assumed
implicit default byte size is 4 bytes due to `extract_limit` in
postgresql adapter did not extract an integer byte size correctly.
That assumption is harmful for other adapters and caused #22594.
In #26386, it is fixed to register integer types limit correctly.
The handling whether out of range or not is only needed with explicit
`limit`, and `Type::BigInteger` is no longer needed.