First argument to form_for must not contain nil
The raised error when passing an array to `form_for` as first argument stated that first argument could not contain nil, when in
fact, it only checks the last element. I've seen rails devs (especially
beginners) stumble across this when building a form for a nested resource.

`form_for [@document, @comment]` if `@document` is nil, this creates the
path `/comments` when the user expected `documents/:id/comments`.

Solved this by raising an error if the array contains any nil values