Fix edge case with `tag` helper
The edge case is that when you would call the `tag` helper with no
content (ex: `tag.p(class: 'hi')`), the `method_missing` would assign
the hash as the content. Now, we are checking if content is a hash or
not, and if it is, we are merging it in with the rest of the options
passed in, which are stored in the `options` hash. We then clean up
after ourselves by setting the content variable equal to `nil`, which is
what it should have been. Kinda hacky, but gets the job done.