Fix edge case with tag helper
The edge case is that when you would call the `tag` helper with no
content and an option hash with string key, the hash would be assigned
as the content. Now, we are using Array#extract_options! to extract
any hash in arguments, and merging it in with the rest of the options
passed in, which are stored in the `options` hash. We then set the
content variable equal to `nil`, which is what it should have been.