Allow the use of listen's 3.1.x branch.
When the initial evented monitor feature was written, the latest version of listen
was the 3.0.x series. Since then the listen project has moved on to the 3.1.x series.
This patch allows the use of the new versions.

This is the 5.0 branch version of https://github.com/rails/rails/pull/26695