Make accessors generated by `Class#class_attribute` thread-safe.
This commit re-implements `Class#class_attribute` to be inherently
thread-safe - without mutexes. The original implementation removed and
re-added methods at runtime which could lead to severe concurrency
issues (i.e. non-deterministic `NoMethodError` exceptions).

This implementation is a drop-in replacement for the existing
functionality and features the same API and functionality as the
original implementation.

Note that this change does not make the use of the method
`class_attribute` itself thread-safe, meaning that calls to this
method should still be performed solely in a single-threaded context
(i.e. during eager loading at boot time). Accessors generated by this
method on the other hand can be used safely in multi-threaded
contexts.