Inflections cleanup
1. Don’t use weird instance variable names
2. Don’t reach into the internals from the test (accessing ivars)
3. Properly reset the inflection rules after each test

(1) and (2) is to establish a clear boundary to make future refactors
easier.

(3) is to fix a bug where we modify the inflection rules and don’t reset
them properly after each test. We only attempt to do that for the `:en`
locale, but we have tests that modifies other locales that doesn’t get
cleaned up properly. For better or worse, the current API for loading
the default rules is to load `inflections.rb`, so let’s just do that and
reset the rules after each test.