Support stop listening to individual broadcasts
Broadcasts can be listened to using stream_for and stream_from but there
isn't anyway to stop listening to individual broadcasts. All broadcasts
can be silenced by using `stop_all_streams`, but that doesn't help the
use case of a client that wants to stop listening to a single broadcast
channel.

This switches @streams from an array of [stream_name, listening proc] to
a Hash with the stream name as the key.  By doing so we can obtain a
reference to the listening proc and use that when unsubscribing from a
named broadcast stream.