Use old typecasting method if no type casted binds are passed in
Query cache doesn't type cast bind parameters since it isn't
actually querying the database, so it can't pass those values in.  Type
casting in the query cache method would cause the values to be type cast
twice in the case that there is a cache miss (since the methods it calls
will type cast *again*).  If logging is disabled, then adding the type
cast code to the query cache method will needlessly typecast the values
(since the only reason those values are type cast is for display in the
logs).

Fixes #26828.