Don't load ActionSystemTest in production
By moving to the TestUnit Railtie, and doing the file requirement
inside the onload call we can avoid loading ActionSystemTest in
production and load it in the test env.

This is important for performance reasons - loading up unnecessary files
and object is expensive, especially when they should never be used in
production.