Catch invalid UTF-8 JSON documents and respond with BadRequest
It's currently possible to send a malformed JSON document on POST that
includes invalid UTF-8 byte sequences. Unfortunately the JSON decoder
will not perform any checks on the validity of the input string and will
include the invalid byte sequence in the deserialized object. This
generally leads to hideous errors further down the stack, since the
sequences are not valid in UTF-8.