Add `Integer#round` and `Float#round` core extensions
`half` option was added to Integer#round and Float#round in Ruby 2.4.
And a behavior of round was changed to round to even.
We have depended a behavior of rounding to nearest.
In some case (i.e. `ActionView::Helpers::DateHelper#distance_of_time_in_words`),
we have to pass `half: :up` to `round`.

This core_ext enables to us to pass a option,
even before Ruby 2.3.