`eager_autoload` needs to be triggered for nested constants.
When `ActiveRecord::Associations.eager_load!` is called, it eager loads
the `Preloader` but does not trigger a `eager_load!` on the `Preloader`
itself. A eager loaded component is responsible to forward the
`eager_load!` call to nested constants that make use of `eager_autoload`
as well.

This will also resolve the occasional `NameError` that we've observed in
production:

```
NameError: uninitialized constant ActiveRecord::Associations::Preloader::BelongsTo
```