Remove teardown code
Since I've moved the teardown code that contains the screenshot handling
to be generated when the application is generated this code was
interfering with the screenshot taking.

Because this runs before any app teardown code we would be resetting
sessions before taking the screenshot, resulting in a blank browser
window. The code to reset the sessions must come AFTER a screenshot has
been taken.