Clear available connections immediately
It'll be re-cleared when it's rebuilt in with_new_connections_blocked's
ensure, but we still need to clear it inside this synchronize -- we've
disconnected connections that may be available in the queue, and while
other threads are not allowed to make *new* connections, they are still
allowed to take existing ones from there.

This was incorrectly removed in d314646c965b045724e6bdb9d61dcecfabc0ba8f.