Introduce `ActiveRecord::ConnectionAdapters::SQLite3Adapter.execute_batch()`
When using a version of SQLite3 that does not support multi-insert statements, we build a string containing
multiple statements. This string is then passed to `SQLite3::Database.execute()` but `.execute()` will only
execute the first statement in a string and return any necessary results.

This change introduces a new method which uses `SQLite3::Database.execute_batch()` to execute all of the
statements in the string. The caveat with `SQLite3::Database.execute_batch()` is that it cannot return a result
set and instead returns nil. See http://www.rubydoc.info/gems/sqlite3/1.3.12/SQLite3%2FDatabase%3Aexecute_batch.

This new method is used to insert multiple versions into the `schema_migrations` table when migrations have been
executed successfully, but only when normal multi-insert statements aren't supported by the SQLite3 version
installed on the system.