Stop using the `pg` Float encoder
PG's type map assumes that all Ruby floats are going to a field with an
OID of type float4 or float8, and generates text which is invalid syntax
for other types. Since the gem can handle floats properly without this
encoder (albeit slightly slower), we can continue to use that as we have
in prior versions of Rails.

Fixes #27246