Cleanup Rails provided helpers
1. Clean up screenshot helper

Updates documentation to be clearer and separates the concerns of saving
the image, setting the image path, and displaying the image.

2. Remove Rails provided assertions for selectors

This was moved upstream to Capybara and is no longer necessary to be
included in Rails

3. Remove form helper

The form helper is pretty specific to Basecamp's needs and may not be
helpful outside of Rails.