Make sure there is a content type when using the deprecated render text
When the accept header is present, the format of the request was being
processed as nil because the Template::Text renderer was setting the
format as a string like `application/json` and this was failing in the
lookup of the types since we use the symbol to do the loopkup.

To fix it we are fallbacking to the format string in case the lookup
returned nil.