Revert "Merge pull request #27098 from sina-s/fix-unexpected-session-sharing"
This reverts commit cbf2e7aef3f3c140955a4612d0f585b81dae5202.

While this bug is a legitimate bug, this method has been broken in this
exact way for quite some time. Any code that is using this method is
almost certainly doing it wrong, as it's pretty much impossible to use
correctly. That code will all be broken by this change.

While this is a good bug fix, and should happen for 5.1, and this method
is only used for tests, we should not break all of the code using it in
a point release.