Disable sorting of params in to_query
The order of query params can be significant. For example in the rails_pjax gem, where the `_pjax` param is stripped from the url using a regex. Because of the sorting the `_pjax` param is always the first, which causes the test not to be representable for a real-world request where query parameters aren't necessarily sorted. 

See rails/pjax_rails#85 for a local fix in the testcase.