Add failing BigDecimal cast test for ruby 2.4
This adds tests to the typecasting of Integer, Float, and Decimal
showing that empty strings should cast to nil, and invalid strings
should result in zero.

All of these tests pass in ruby 2.3.3, and all earlier versions. The
tests against BigDecimal fail on ruby 2.4.0-rc1.

This happens because ruby changed the behaviour of BigDecimal(str) to
better match methods Integer() and Float(), which return errors.

See also:
https://bugs.ruby-lang.org/issues/10286
https://github.com/ruby/bigdecimal/commit/3081a627cebdc1fc119425c7a9f009dbb6bec8e8