Move DATABASE_URL merge logic to config resolver
With the `DATABASE_URL` logic living in the connection config resolver, we don't
need to mess around with enviroment(such as RAILS_ENV, RACK_ENV) in it.
Because Active Record will now use the `local_configurations`, which is a
configuration hash for the running environment, the `DATABASE_URL` logic
gets easier to handle as it only needs to set the `url` config for the
`primary` database config, which is the default one.