make backburner integration test to work
Currently, backburner integration test is not running on CI.
https://travis-ci.org/rails/rails/jobs/196005322#L610

Using `Backburner::Worker.connection` to check whether beanstalkd is
running. But `Backburner::Worker.connection` was removed in backburner
1.2.0.
https://github.com/nesquena/backburner/commit/81fde499c2263abe8dedee59a3d0e03e92d72627

Therefore, this check process always becomes false, so the test is no
longer done. I fixed it so that check processing is done correctly.