Properly nest core classes under a "Core Extensions" label
Since Active Support is monkey patching a lot of core classes, let's
rather document these changes under a new section so they are still
documented but not encumbering the sidebar.

We can safely remove the rescuing of the `LoadError` since as of
cd7cc525, it's not possible to generate the API from an application.

[ci skip]

[Kasper Timm Hansen & Robin Dupret]