Do not use `Arel.star` when `ignored_columns`
If there are any ignored columns, we will now list out all columns we
want to be returned from the database.

Includes a regression test.

Re: the test I had to change, it should have never been passing in the
first place, due to the ignored `first_name` column on `Developer.