Remove version number from API links when coming from Edge Guides
Rails Guides links to the API documentation include a version number matching
the Rails Guide providing the link. However, the 'version number' for Edge
Guides is a git commit hash, not a true version number, which results in a
404 when you try to follow the documentation link.

This change fixes that problem by only including 'true' version numbers, and
also removes some unused code.