Deprecate calling `attr_will_change!` with non-attributes
This was never really intended to work (at least not without calling
`define_attribute_methods`, which is less common with Active Record). As
we move forward the intention is to require the use of `attribute` over
`attr_accessor` for more complex model behavior both on Active Record
and Active Model, so this behavior is deprecated.

Fixes #27956.
Close #27963.

[Alex Serban & Sean Griffin]