Fix test broken after config rename.
When renaming the secrets config to read_encrypted_secrets.

The frameworks tests break because the remove the config/environments
files, which included production.rb that had the flag enabled.

As such some of the tests couldn't find a secret_key_base.

They also couldn't a key base because the we no longer generate a secret
key base in the production environment in secrets.yml.

Fix this by assigning a key base to our config.