Include selects in group query with having clause
When a grouped calculation contains a having clause that references a
selected value, we need to include that selected value in the query.

Postgres doesn't support referencing a selected value in a having
clause, but other databases do; we can skip the test on the pg adapter
but run it for the others.

This was fixed before in 9a298a1, but the test coverage was lost in
5a05207. The fix regressed in 6311975 and was removed in 97d46c1.

(cherry picked from commit 18125683ac2a23971b64a79d256ba6e252304c6f)