Make required by default test for belongs_to association clearer
Since #18937 `belongs_to` associations receive a setting to determine if
it should be or not treated as `required` by default.

While the tests were still passing, it was not evident that the
"default" behaviour for `required` could change in fuction of a setting,
that is set by default for fresh Rails5 apps, but not for upgraded
apps.

This commit try to relate them to make it clear what is the behaviour
expected when the setting is set as `true` or not set.