Properly check transaction in persistence
```
[NoMethodError]: undefined method `state' for nil:NilClass  Method:[rescue in block in refresh]
```

In `within_new_transaction`, there is the possibility that
`begin_transaction` returns a `nil`. (i.e.: so `transaction = nil`)
So this method is checking `transaction` for nil in 2 spots.

Unfortunately, there is one line that is not checking `transaction` for `nil`
That line, `commit_transaction`, throws an exception for us in AR 5.0.0.1

The problem with the method is finally realized in the error checking itself.
it calls `transaction.state` (i.e.: nil.state) and that is the final exception
raised.

The actual underlying (user) issue is hidden by this line.

Solution is test transaction for nil.