Introduce ActiveSupport::TestCase.describe for grouping related tests
We in the Ruby world, write lot's of tests. And this is awesome! 🙌

Now writing lot's of tests means we do have to manage them somehow. See
my controller tests, taken right from an app I've been working on.

This is RESTful controller here, that is something a bit more specified:

```ruby
class ParentsControllerTest < ActionDispatch::IntegrationTest
  test 'POST /v1/parents registers a new parent'
  test 'POST /v1/parents allows empty email for the kid'
  test 'POST /v1/parents validates timezone a new parent'
  test 'POST /v1/parents respond with errors'
  test 'POST /v1/parents does not allow duplicate email addresses'

  test 'PUT /v1/parents/:id updates a parent'
  test 'PUT /v1/parents/:id can set a card'
  test 'PUT /v1/parents/:id can update card expiry through :credit_card'
  test 'PUT /v1/parents/:id can update card expiry through :card'
  test 'PUT /v1/parents/:id card_token takes priority over {credit_,}card'
  test 'PUT /v1/parents/:id can set a new email'
  test 'PUT /v1/parents/:id can set a new password'
  test 'PUT /v1/parents/:id returns a new token'
  test 'PUT /v1/parents/:id errors out on empty password'
  test 'PUT /v1/parents/:id can set a new password and change email at the same time'
  test 'PUT /v1/parents/:id works only for the parent in the token'
  test 'PUT /v1/parents/:id can error out with invalid stripe token'

  test 'GET /v1/parents returns the partners of a parent'
end
```

And this is a PORO test, which can represent something a bit more
generalized:

```ruby
class AuthorizableTest < ActiveSupport::TestCase
  test '#confirm does not confirm a nil input'
  test '#confirm proper token and flags confirmed_at'
  test '#password? is falsy on for blank password_digest'
  test '#password? is truthy on for present password_digest'
  test '#initial_password sets the initial password with confirmation token'
  test '#initial_password with wrong confirmation token'
  test '#initial_password does nothing for already set password'
  test '#reset_password resets a password with correct code'
  test '#reset_password with wrong code'
  test '#reset_password with unconfirmed auth'
  test '#reset_password with blank code'
  test '#reset_password attack for blank password_reset_code'
  test '#new_password_reset_code generates a random 6 digit code'
  test '#new_password_reset_code works for nil passwords'
  test '#change_password can change the password if old_password matches'
  test '#change_password cannot change empty password'
  test '#change_password cannot change empty short passwords'
  test 'password cannot be set or kept to nil after creation'
end
```

When you unroll the tests, it doesn't get that hard to navigate
through , but still, it's a bit noisy. We may argue that this way
structuring tests is not the optimal one. I think, I'm quite the average
testing guy here and lot's of folks tend to converge to this
organizational style. Again, or so I think. 😅

Now, if we had a way to group those tests by something, even just
organizationally, it would be a bit easier for me to navigate, know
where to put another `#reset_password` test or fold the tests I don't
need to focus on right now in my editor.

Here is the controller:

```ruby
class ParentsControllerTest < ActionDispatch::IntegrationTest
  describe 'POST /v1/parents' do
    test 'registers a new parent'
    test 'allows empty email for the kid'
    test 'validates timezone a new parent'
    test 'respond with errors'
    test 'does not allow duplicate email addresses'
  end

  describe 'PUT /v1/parents/:id' do
    test 'updates a parent'
    test 'can set a card'
    test 'can update card expiry through :credit_card'
    test 'can update card expiry through :card'
    test 'card_token takes priority over {credit_,}card'
    test 'can set a new email'
    test 'can set a new password'
    test 'returns a new token'
    test 'errors out on empty password'
    test 'can set a new password and change email at the same time'
    test 'works only for the parent in the token'
    test 'can error out with invalid stripe token'
  end

  describe 'GET /v1/parents' do
    test 'returns the partners of a parent'
  end
end
```

And here is the PORO:

```ruby
class AuthorizableTest < ActiveSupport::TestCase
  describe '#confirm' do
    test 'does not confirm a nil input'
    test 'sets proper token and flags confirmed_at'
  end

  describe '#password?' do
    test 'falsy on for blank password_digest'
    test 'truthy on for present password_digest'
  end

  describe '#initial_password' do
    test 'sets the initial password with confirmation token'
    test 'wrong confirmation token'
    test 'does nothing for already set password'
  end

  describe '#reset_password' do
    test 'resets a password with correct code'
    test 'with wrong code'
    test 'with unconfirmed auth'
    test 'with blank code'
    test 'attack for blank password_reset_code'
  end

  describe '#new_password_reset_code' do
    test 'generates a random 6 digit code'
    test 'works for nil passwords'
  end

  describe '#change_password' do
    test 'can change the password if old_password matches'
    test 'cannot change empty password'
    test 'cannot change empty short passwords'
  end

  test 'password cannot be set or kept to nil after creation'
end
```

I know David doesn't prefer the BDD DSL, but I wanna make a case for
`.describe` here. It doesn't create hidden test-cases or special
environments. It cannot be nested by design, so we don't promote those
levels into levels into levels kind of tests. It's used purely for
organizational work.

There are lots of places in the Rails tests where we can benefit from
grouping tests together that way as well, but I think it will shine for
applications. We may change the method name, but I think the ring with
the existing terminology may help in people discovering it.