Sync transaction state when accessing primary key
If a record is modified inside a transaction, it must check the outcome
of that transaction before accessing any state which would no longer be
valid if it was rolled back.

For example, consider a new record that was saved inside a transaction
which was later rolled back: it should be restored to its previous state
so that saving it again inserts a new row into the database instead of
trying to update a row that no longer exists.

The `id` and `id=` methods defined on the PrimaryKey module implement
this correctly, but they are almost always overridden by dynamically
defined attribute methods. Those attribute methods must also check the
transaction state if they are the record's primary key.