Use current database in MySQL's `table_exists?`
This is a minor regression from when we switched from using SHOW TABLES
to querying the information_schema for `table_exists?`. SHOW TABLES by
default querys the currently selected database but the current
implementation defaults to `@config[:database]` if no database name is
prepended to the table name. This meant that if you issued a `USE`
statement to change database, `table_exists?` gives spurious results as
it's still querying `@config[:database]`.

The solution is to use the MySQL `DATABASE()` function in place of
`@config[:database]`.