[ci skip] Add console examples for each association type.
For kinesthetic learners (doers) it is important to show concrete
examples of how advanced concepts like associations work. What I
have attempted to do here is (using the console)
give examples of how each of the associations works based on the models
and attributes for each. Existing text/examples were left in place and
supplemented with
console examples. Note that I did remove the published_at attribute
in 2.1 belongs_to and 2.3 has_many and replace with :name as it seems to
make more sense
for the examples I am using. I also updated the diagrams/images that
went along with these
examples.