Keep frozen records frozen on save
In Rails 4.2.6, if you called `save` on a frozen record, it would keep
its frozen state. Between 4.2.6 and 5.0.0,
07723c23a7dc570beae73c074ad37227e3e8a06e introduced a regression where
`save` will mutate a record and unfreeze it.

I pinpointed it to ActiveRecord::AttributeMethods::Dirty's
`changes_internally_applied` that replaces a record's
`@mutations_before_last_save` and `@mutations_from_database` ivars.

Returning if the record is frozen bypasses the changes and keeps the
records frozen.