Correct explanation of strong parameters
This change corrects a misunderstanding of the way required parameters
are handled. Whilst there is an exception wrapper that maps
ActionController::ParameterMissing to a 400 bad request, this is only
used in logging. The current description implies that the exception is
caught and a 400 Bad Request response is sent. This is not the case.