Adding App to load path for command lookup
Moving the `.root` method to the Rails::Command::Bevavior module so it
can be used by both generators and commands. Also using `.root` method
to add either a Rails engine or app to the load path when available.
This will provide support for builders of Rails engines and apps to add
their own Thor commands and have them be accessible via the `rails`
command.